requirements.txt
google-generativeai>=0.8.0
mem0ai>=0.1.0
pyautogui>=0.9.54
pynput>=1.7.6
mss>=9.0.1
Pillow>=10.0.0
faster-whisper>=1.0.0
elevenlabs>=1.0.0
pyttsx3>=2.90
opencv-python-headless==4.8.1.78
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
python-dotenv>=1.0.0
pyperclip>=1.8.2
gTTS>=2.5.0
av==13.1.0
duckduckgo-search>=6.0.0
beautifulsoup4>=4.12.0
requests>=2.30.0
